changed the text file names for the logs (not sure if SimulatorLogs.log is present or not).
fixed it so that the configurator logs dont print to the console and
just print to the file.

made the logger for the simulator and fixed the logging lines, still
need to put an instance of the LoggingFrame class witht he simulator log
file ( wasn't sure where to put it) since it needs to be put after the
window is closed. if you have an idea on where to put it just put there:

LoggerFrame simLog = new LoggerFrame("Simulator
Logs","SimulatorLogs.log")